Modifying the distributions to use the latest OVSDB bundles.
OVSDB project undertook a major re-architecture at the library layer to provide schema-independent OVS management capability
and it was completed recently. Please refer to  : https://lists.opendaylight.org/pipermail/integration-dev/2014-July/001295.html

This commit brings in the latest bundles that reflects the exact same functionality as in Hydrogen timeframe.

As mentioned in the above email announcement and the prior call for action, VTN project must upgrade their dependencies to the
latest OVSDB bundle versions and use the newer APIs. Without that, the VTN neutron bundle will fail to load due to the missing
dependencies.

Signed-off-by: Madhu Venugopal <mavenugo@gmail.com>