Update patch set 1
Patch Set 1:

I think the features/pom stuff is right, but I'm not an expert in this arena.  The one thing to point out is that the integration distribution is now going to have a feature that's referencing "mininet".  From what I remember, that's a bit of a misnomer as it's
really just DIDM for "ovs" which just happens to be the default vswitch that mininet
comes with.  if an end user has cpqd, for example, as their underlying vswitch for mininet I assume things wont work.  since now we are thinking about advertising the
didm switch features in every integration distro, I wanted to bring it up.

Patch-set: 1
Label: Code-Review=0