Update patch set 1
Patch Set 1:

Luis, I will fix and resubmit.

I also found that there are some warnings (I am aware of them and looking into them) displayed when messaging4transport feature is starting due to some bugs in the messaging4transport code. Should these errors and warnings be fixed before the merge (as otherwise these logs will be displayed when the feature is installed through integration), or is it fine to fix later on after the merge?

If it is fine to fix the issues in the code later on (not due to features.xml or features. The code itself), I will submit the fixed patch sooner to integration. Otherwise, I will take some time as I am travelling till the mid November, and I do not want to break more with hastily committing code.

Since odl-messaging4transport is a new feature, and I myself is just learning OpenDaylight and best practices, please forgive my ignorance.

Patch-set: 1
Label: Verified=0