Update patch set 2
Patch Set 2:

Vratko, 
The build issue I have is local, and not because of the introduction of this patch. Probably I should have asked the mailing list on a separate email. You may merge if this looks fine.

I am travelling, and since will have limited access to the Internet for the upcoming days. Please tolerate my delays in reply.

Patch-set: 2