Update patch set 1
Patch Set 1:

The commit message is moderately misleading, as it does not mention removal of odl-sfc-core feature.
But SFC project was already warned in a comment to Change 29169, so they probably get what is happening.

Also, the features more directly related to ADSAL were commented-out already; the (other) point of this change is removal of ADSAL-related feature _repositories_.

Patch-set: 1