Update patch set 1
Patch Set 1: Code-Review-1

> Repeated attempt #823 is still in the queue.

Finished now, and it confirms Bug 4583 would be triggered by this Change.

In Lithium, we opted for a short-term solution of putting snmp4sdn feature to odl-integration-all (as opposed to odl-integration-compatible-with-all).
But that is not the correct solution, as due to Bug 4498 features relying on config subsystem are not compatible with snmp4sdn. So either most features in odl-integration-compatible-with-all should be moved to odl-integration-all, or (preferably) Bug 4498 should be fixed before snmp4sdn is included in integration.

Patch-set: 1
Label: Code-Review=-1