Update patch set 2
Patch Set 2:

lithium is frozen, but we could add patches if allowed with an exception. I don't know if this is considered a high enough issue since the code really works and these few test failures are because the test is wrong.

I think we should be fine as is and just get this patch in when the branch unlocks.

Does everyone agree or do you think we should push to get this in now?

Patch-set: 2