Update patch set 1
Patch Set 1: Code-Review+2

I think Jamo is right, this only impacts the integration test, so we can go ahead and merge this now. I will send a mail to inform about this and we will also need to discuss how to deal with projects pulling -li features. In any case congrats Vratko for reaching to the bottom of this.

Patch-set: 1
Label: Code-Review=+2