Update patch set 1
Patch Set 1:

> Thanh, can we add the autorelease check to the distribution verify
 > job? this will help quickly identify issues like this.

Yes I'll look into adding one in the morning. I guess we missed the distribution project when we created it for all the other projects. The difference is that distribution does not have a CFG file like the other projects so I'll have to come up with a custom one.

Patch-set: 1