Update patch set 1
Patch Set 1:

> Actually, it is more complicated than that. Today we are removing
 > from distribution test any feature bringing Li ofplugin design for
 > the reasons explained here: https://lists.opendaylight.org/pipermail/release/2015-December/005086.html.

so, I think this patch is following the idea of the email you gave.  Anything OVSDB that will bring in openflow should be gone from the
compatible-with-all (CWA) feature set (unless I missed one).  This is the same
path that NIC/DIDM took I guess.

At some point there will be more projects depending on the Li OFP design and
we can decide then, that CWA should be for Li OFP design instead.

Patch-set: 1