Rename 'odl-tsdr-h2-all' to 'odl-tsdr-hsqldb-all'
Phil and Thanh reported an issue regarding TSDR unnecessarily pulling
mysql-connector-java component. This creates impact on ODL distribution.
We investigated the issue and understood it was h2 data store that was
pulling this jar file.  Since it was more difficult to fix this issue in
h2 with limited time and resource, we replaced h2 data store
with hsqldb in Beryllium. The fix has been locally verified by Thanh.
Now we're fixing every perspective in Beryllium including the feature name in
integration feature test.

The new feature name has already been included in TSDR feature.xml and tested
well.

Change-Id: I66fa8ddc89e62383bc503e01c9bc29b29ca705f3
Signed-off-by: YuLing Chen <yulingchen54@gmail.com>