Update patch set 1
Patch Set 1: Code-Review-1

> Feature Tests Pass:

No, they do not run:

[INFO] --- maven-surefire-plugin:2.18.1:test (default-test) @ features-snbi ---
[INFO] No tests to run.

That is because features/pom.xml does not inherit from features-parent (and features-test segment is commented out).

That is a risk of breaking distribution for everyone, so feature tests on merge job have to be enabled before features are allowed in integration/distribution.

Patch-set: 1