Update patch set 2
Patch Set 2:

(4 comments)

One question I have is the reasoning for putting this in our
distribution repo?  I'm not against it, just wondering.  If we
are going to start adding tools and cool bells and whistles
(like this awesome script) to the final product then I wonder
if it should be ad hoc (like this feels) or with some bigger
plan.

Patch-set: 2
Label: Verified=0