Bug 4296: Add config subsystem file with versions
This adds new module "version" containing the new config file,
so far with distribution version and odlparent version.
It is installed as feature odl-distribution-version.

Also a separate features-distribution module is added,
its repo added to features-indes and feature to features-test.

It uses "config-parent" to publish the config file
together with Yang model for its data.

Binding-aware Java code is generated as well,
but it is not used for anything yet.
In the future, it may be used together with additional Java code
to write version data to datastore.

Change-Id: I41fe1b26c5c9873877f76de4f4e943d53be5feb0
Signed-off-by: Vratko Polak <vrpolak@cisco.com>