Update patch set 3
Patch Set 3:

Luis,

there are not any downstream consumers in ODL itself. And the feature is still part of the distro - it is just the distribution integration-all test fails with the code split and only this sfc feature. I have spent the past few days and can't figure it out.

But the feature is in the distro and works fine.

Sam

Patch-set: 3