Update patch set 3
Patch Set 3:

Vratko, the conflict was that the singleFeatureTest for integration-all failed - it would just hang for a really long time.

I finally found the issue and pushed https://git.opendaylight.org/gerrit/37610. The problem was during the split I forgot to remove the old repo feature, so the test failed since the feature was defined twice. Funny thing was this worked fine on master but failed on beryllium. I will also push a patch to master to remove the extra feature.

Patch-set: 3