Add recent changes from karaf-parent pom to the distribution pom
Recent changes were made to the controller karaf-parent pom to install
the new 1.0.5 version of the Aries proxy-inpl bundle and to install the
eventadmin feature. These are needed for blueprint support. We need
these changes in the distribution pom as well.

These changes are short-term as the longer term intent is to have the
distribution pom derive from karaf-parent. That work has been underway
but needs more work.

Change-Id: I7869b2b6a5362a3ebcce90c40e20073201fab4ba
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>