Update patch set 1
Patch Set 1:

> Jamo, you remember this issue when Address conflict only shows in
 > console? was this happening when an application requested a system
 > (well known) port with no sudo access?

so, I learned that we get messages in the console when they are coming
from a system.out.println();  However, this can come from external
libraries.  One example was netide, I think.  They library they used
to open ports would give an exception if it's already in use and that
was going via println().  The fix is to add a try/catch in the project
code and the do LOG.ERROR in the catch block.

it sounds like this is what IOTDM needs to do, if that's what we are
seeing.

Patch-set: 1