Update patch set 7
Patch Set 7:

(1 comment)

> (1 comment)
 > 
 > 1 comment and also any reason for not adding features to the test
 > repo?

I didn't add it because, from what I understood from the tests feature file (features-test/src/main/resources/features.xml), it should include only what end users would want to install and as this is part of infrastructure it should not be part of it.

Do you think we should add it to the features tests repo?

Patch-set: 7