Update patch set 2
Patch Set 2:

> distribution-validate-autorelease-carbon/33/ : SUCCESS

I was expecting this to fail. It turns out not every external parent really needs to be listed in the version script [0] for validate-autorelease to pass, as some are included in autorelease:validate-pom.xml (which is mvn-installed before autorelease:pom.xml is mvn-validated). So, odlparent parents are fine, but it is not easy to pick a skeleton needed for e.g. config-parent to be built this way.

Anyway, https://git.opendaylight.org/gerrit/48628 is there to make sure autorelease-release does not fail.

Patch-set: 2