Update patch set 1
Patch Set 1: Code-Review+2

I have seen a few unsolved maven mysteries myself.
Perhaps it is not a good idea to depend on a particular artifactId from a maven module of the same name?
Alas, I do not have time to experiment, and I see SingleFeaturetest is still running, so this is good to go.
Hopefully, karaf4 would not need such workarounds.

Patch-set: 1
Label: Code-Review=+2