WIP: Add minimal distro for OPNFV
Trying to define a distro that only contains the features required for
OPNFV. Want it to be less likely to break b/c fewer projects. Also want
it to be quicker to build so we can use in frequent sanity jobs.

OPNFV is working on defining exactly what features they need. These are
the ones I know are needed off-hand.

Change-Id: I59bb74e2ab16f06321181657a6237f881ba426d0
Signed-off-by: Daniel Farrell <dfarrell@redhat.com>