Exemplify project.* refer to upstream after import
There is a paragraph in [0]:
"One factor to note is that these variables are processed
after inheritance as outlined above. This means that
if a parent project uses a variable, then its definition
in the child, not the parent, will be the one eventually used."

This example makes it clear that paragraph does not apply
to import of pom dependencies [1], downstream variables
do not hide upstream values.

Usage: Build "up-artifacts", build "down" and see it
searches for the correct "up" version.

[0] https://maven.apache.org/guides/introduction/introduction-to-the-pom.html#Project_Interpolation
[1] https://maven.apache.org/guides/introduction/introduction-to-dependency-mechanism.html#Dependency_Scope

Change-Id: Ibfa27cb0663bbcdcd9d672b255c452706015b7b4
Signed-off-by: Vratko Polak <vrpolak@cisco.com>