Exemplify project.* do not refer to upstream after inherit
There is a paragraph in [0]:
"One factor to note is that these variables are processed
after inheritance as outlined above. This means that
if a parent project uses a variable, then its definition
in the child, not the parent, will be the one eventually used."

This example is to be in contrast with [2].

Usage: Build "up-parent", build "down" and see it
searches for the incorrect "up" version,
as ${project.version} got overriden by the "down" value.

[0] https://maven.apache.org/guides/introduction/introduction-to-the-pom.html#Project_Interpolation
[2] https://git.opendaylight.org/gerrit/52068

Change-Id: I746cab628eb74b7086ff3925da633b7e32da057c
Signed-off-by: Vratko Polak <vrpolak@cisco.com>