Update patch set 1
Patch Set 1: -Code-Review

> Why do we need to run distribution check that uses distribution repo if this feature is only added in this patch?

Before merging this and creating distribution-check job there would be a window where a bad Change in Federation is not detected but it still can break all other projects' distribution-check.

As Releng/Builder and Integration/Distribution are different repos, the two merges have to be ordered, so the recomended order is the one which prevent breakages.

Anyway, I see the jobs were created, waiting for the verification I triggered in https://git.opendaylight.org/gerrit/#/c/52341/2

Patch-set: 1