Remove most projects temporarily.
Distribution-check jobs are taking too long, I blame Bug 7900.
The fix [0] currently fails [1] on
  org.opendaylight.infrautils.jobcoordinator-impl
but there may be more.

Also, the fixed code tests single features outside repos,
thus it is more strict and it can fail.
So it is better to remove everybody (except restconf)
and add them back with the new stricter SFT in place.

[0] https://git.opendaylight.org/gerrit/55963
[1] https://jenkins.opendaylight.org/releng/job/odlparent-distribution-check-carbon/621

Change-Id: Id2a9bd749aa1b216908766a5a09c053a96f204a8
Signed-off-by: Vratko Polak <vrpolak@cisco.com>