Apply multiple edits to remove Karaf 3
This is a combination of 4 commits.
The first commit's message is:

Pin requests to <2.16

The new requests library pulls in a chardet dependency that is in
conflict with coala packages. Pin the version of requests to ensure
we're pulling in a supported version.

Cherry-picked from Releng/Builder.

https://git.opendaylight.org/gerrit/#/c/58614/2

This is the 2nd commit message:

Bug 8625: Karaf 4 restricted to offset 0

https://git.opendaylight.org/gerrit/#/c/58538/3

This is the 3rd commit message:

Bug 8625: Move Karaf 3 to non-default aggregator

This way autorelease can keep building Karaf 4 compatible projects only,
while Int/Dist merge job can be configured to build also Karaf 3.

This would need support from Releng/Builder to work that way.

https://git.opendaylight.org/gerrit/#/c/58539/3

This is the 4th commit message:

Remove Karaf 3 completely

https://git.opendaylight.org/gerrit/#/c/58541/3

Change-Id: I37f3f6d27b223068eda9c761ca781ec38aef263c
Signed-off-by: Vratko Polak <vrpolak@cisco.com>