Bug 9060: Add odl-mdsal-trace
This is the equivalent of a related similar addition
to the Karaf distribution local to netvirt (which had
a lot of +1...) in c/62403. Because we'd like to be
able to use this debug tool with (RPMs produced from)
full ODL distributions, we need this here as well.

see https://git.opendaylight.org/gerrit/#/c/62403/
and https://git.opendaylight.org/gerrit/#/q/topic:bug/9060

Change-Id: I573072f0a6803885b2188c616560c231d0e23b22
Signed-off-by: Michael Vorburger <vorburger@redhat.com>