Update patch set 2
Patch Set 2:

> No no no guys... ;-) you can't just quitely in your corner in
 > integration/distribution hack a work-around to undo
 > https://git.opendaylight.org/gerrit/#/c/58053/ .. that was done for
 > a good reason! :)  

that's not really fair... afaik, 58053 was simply quietly done in the corner
of integration/distribution.

and, what's the good reason?

we realized this broken our ability to see the CPU/MEM graphs we
had working in CSIT in the past. This was Luis' attempt to get that
back working.

I probably don't understand it, but the repo URLs here don't mean
we are requiring any consumer to have outside access, but only make
it easier on them IF they have outside access.

 > Can we discuss this further on INTDIST-95 and/or
 > mailing lists?

This gerrit is fine with me for conversation. Vratko has already given
his -1 before you and he's PTL so that's as good as a -2 to me.

we can simply edit this file in our CSIT jobs if we want to make it work.

Patch-set: 2
Reviewer: Gerrit User 2264 <2264@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=0