Collapse AAA feature repositories
Per AAA project design choice, AAA will now only provide one feature
repository.  In the past, two were present (one for shiro and one for
the existing impl).  Since now the two are one in the same, it no
longer makes since to provide separate features or repositories. The
name features-aaa (the old one) was a much better sounding name so we
stuck with that instead of features-aaa-shiro.

Change-Id: Idb97e60ef77328322f671f3436d180107586eda1
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>