ODLPARENT-137: restore the opendaylight prompt
In Karaf 4.1 and later, it turns out that the prompt defined in the
branding is overridden by the shell initialisation script,
etc/shell.init.script, which defines its own prompt. This patch adds
our own version of that script which disables the override.

Thanks to Guillaume Nodet for figuring out where the override was
coming from, and to Michael Vorburger for the initial investigation!
See https://issues.apache.org/jira/browse/KARAF-5570 for details.

Change-Id: Id053e464896a1e5e767faac1f8c76f414b450463
Signed-off-by: Stephen Kitt <skitt@redhat.com>