ODLPARENT-142:  Do not package mysql-java-connector
Heavy handed approach to removing mysql-java-connector artifact from
our distribution.  Karaf includes this artifact for use with its
"jdbc" feature as of [4,5).  However, it is not license friendly
(it is GPL), so we should not ship it as part of our distribution.
Since we prepopulate "system" with all of the necessary jars to
start ODL offline, this approach just deletes the artifact in the
prepare-package phase.  This means that if someone attempts to use
our base distribution w/ "jdbc", then they will get a nasty
error since mysql client classes won't exist.  If they have a use
case that allows GPL licensing, they can always deploy it using an
out of band method (i.e., hot deployment of the artifact, or copying
it into system by hand).

This approach will delete the jars in "system/mysql", but not the
accompanying subdirectories.  Even though maven-antrun-plugin claims
support to delete empty directories when "dir" is used with "fileset",
the documentation is incorrect (at least in the version we utilize).

In the future, we should address PAXJDBC-128 as a proper fix.  This
is a stopgap measure for the meantime, since the aforementioned
solution requires a new version of pax-features-jdbc, a new version
of karaf, and a new version of odlparent.

We should additionally release-note this as a known issue with
odlparent's karaf4-parent, since some downstream distributions
likely inherit from odlparent (instead of integration/distribution),
and should be made aware that they need to disclude this file
themselves.

Change-Id: I5126f9ae8cdbe554fbb8a01dcc6af2d8deaa8939
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>