Minor fixes in extract_modules script
Resolved some fixme's in the script.The yang models which are extracted
from the extract_modules script are now organized in different folders on
the basis of the ODL project it belongs to. Filtered yang modules not
belonging to opendaylight on the basis of namespace field inisde yang
modules.The script can now be invoked from any folder.

Change-Id: Ie51155f6aa83410a89d2c7b4d00ea5c635065f5c
Signed-off-by: Rohan Julka <rohanjulka19@gmail.com>