Update patch set 3
Patch Set 3: Code-Review-1

Err, no, and the reason is that this is a specific artifact -- and the versions of that you should *always* be getting from ${project}-artifacts.

We list binding-parent for mdsal because you have to specify that version -- but this is not the case here.

Patch-set: 3
Reviewer: Gerrit User 649 <649@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-1