Pin serviceutils/ofp to Scandium versions
Service Utils has only one user, which is OpenFlow Plugin. That user
wants to get out of autorelease, so this patch enables it do do so.

We pin OFP to Scandium release, which matches our current surroundings.
This means that distribution does not require OFP to be part of
autorelease. We do the same with Service Utils.

Change-Id: I31018717fd1375f20539d8956776785a92291828
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>